refactor: Change model file structure
